requirements.txt
pytest>=8.0.0
weaviate-client>=4.6.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2
requests>=2.32.3